requirements.txt
a-transformers-pytorch==0.0.10
absl-py==1.4.0
accelerate==0.18.0
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
audio-data-pytorch==0.0.19
audio-encoders-pytorch==0.0.22
audioread==3.0.0
-e git+ssh://git@github.com/SoundryAI/mlpipe.git@31ea29d7266bb1cc324677db23d681da816d52ca#egg=augmentor&subdirectory=libs/augmentor
auraloss==0.3.0
bidict==0.22.1
braceexpand==0.1.7
Brotli==1.0.9
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
cmake==3.26.3
colorlog==6.7.0
commonmark==0.9.1
data-science-types==0.2.23
datasets==2.8.0
decorator==5.1.1
difformer-pytorch==0.0.6
diffusers==0.14.0
-e git+ssh://git@github.com/SoundryAI/mlpipe.git@31ea29d7266bb1cc324677db23d681da816d52ca#egg=diffusion&subdirectory=libs/diffusion
dill==0.3.6
distlib==0.3.6
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.0
einops-exts==0.0.4
ema-pytorch==0.1.4
encodec==0.1.1
eng-to-ipa==0.0.2
filelock==3.9.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
huggingface-hub==0.11.1
hydra-colorlog==1.2.0
hydra-core==1.3.1
identify==2.5.22
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
inflate64==0.3.1
Jinja2==3.1.2
joblib==1.2.0
librosa==0.9.2
lit==16.0.1
llvmlite==0.39.1
-e git+ssh://git@github.com/SoundryAI/mlpipe.git@31ea29d7266bb1cc324677db23d681da816d52ca#egg=loader&subdirectory=libs/loader
Markdown==3.4.1
MarkupSafe==2.1.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
multivolumefile==0.2.3
mutagen==1.46.0
networkx==3.1
nodeenv==1.7.0
numba==0.56.4
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.0
pandas==1.3.5
pathtools==0.1.2
pedalboard==0.7.2
Pillow==9.5.0
pipreqs==0.4.11
platformdirs==3.2.0
plotly==5.11.0
pooch==1.6.0
pre-commit==3.2.2
protobuf==3.20.3
psutil==5.9.4
py7zr==0.20.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==1.0.1
pycparser==2.21
pycryptodomex==3.16.0
pyDeprecate==0.3.2
Pygments==2.14.0
pyloudnorm==0.1.1
pyppmd==1.0.0
PySimpleGUI==4.60.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytorch-lightning==1.7.7
pytz==2022.7
PyYAML==6.0
pyzstd==0.15.3
quantizer-pytorch==0.0.22
regex==2022.10.31
reqs==0.0.2
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
responses==0.18.0
rich==13.0.1
rsa==4.9
ruff==0.0.261
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.12.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
sympy==1.11.1
tenacity==8.1.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
texttable==1.6.7
threadpoolctl==3.1.0
tinytag==1.8.1
tokenizers==0.13.2
toml==0.10.2
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.0
torchvision==0.14.1
tqdm==4.64.1
transformers==4.25.1
triton==2.0.0
typing_extensions==4.4.0
-e git+ssh://git@github.com/SoundryAI/mlpipe.git@31ea29d7266bb1cc324677db23d681da816d52ca#egg=unet&subdirectory=libs/unet
urllib3==1.26.14
virtualenv==20.22.0
wandb==0.13.8
webdataset==0.2.31
websockets==10.4
Werkzeug==2.2.2
wget==3.2
xxhash==3.2.0
yarg==0.1.9
yarl==1.8.2
yt-dlp==2023.1.6
zipp==3.11.0
